Add logdie if subcommand fails
Previously the output of a command was logged but a failure in a qx subcommand
did not terminate the pipeline.
This is now changed.
run_command can still be called with 1 as third parameter (ignore_error)
to account for special cases like Gblocks which has exit code 1 even on success.

Fixes: #9