Remove -b1 and -b2 parameters from Gblocks command
The b1 and b2 flags were meant to explicitly set the default value.
This was not the case as the default is a percent value while the parameters set absolute values.
Therefore the following warnings were produced:

WARNING: Parameter -b1 not properly entered.
The minimum number of sequences for a conserved position must be lower than or equal to the minimum number of sequences for a flank position

WARNING: Parameter -b2 not properly entered.
The minimum number of sequences for a flank position must be lower than or equal to 5 (the number of sequences)

This can be avoided by not setting -b1 and -b2 manually.
It also leads to strange cutoffs if the number of sequences in the alignment is far higher than 50 or 85.